refactor: 수경 님 코드와 합침
