+++ src/types/authClient.types.ts
@@ -6,25 +6,10 @@ export interface ISignUpDataLocal {
   hashedPassword: string;
 }
 
-// 일반 회원가입 자료 구조 : FE에서 넘어옴
-export interface ISignUpRequest {
-  name: string;
-  email: string;
-  phoneNumber: string;
-  password: string;
-  passwordConfirmation: string;
-}
-
 // 일반 로그인 자료 구조 : DB
 export interface ILoginDataLocal {
   email: string;
   hashedPassword: string;
-}
-
-// 일반 로그인 자료 구조 : FE
-export interface ILoginRequest {
-  email: string;
-  password: string;
 }
 
 // 토큰 자료 구조
